+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+## 2021-03-11
+- feature: added dependencies to firebase and analytics plugin (https://outsystemsrd.atlassian.net/browse/RMET-547)
+
 ## 2021-03-09
 - feature: added ci pipeline (https://outsystemsrd.atlassian.net/browse/RMET-546)
 - feat: created javascript layer (https://outsystemsrd.atlassian.net/browse/RMET-548)